refactor: 예약 객체 분리
