+++ src/main/java/roomescape/Reservation.java
@@ -1,0 +1,16 @@
+package roomescape;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record Reservation(
+    Long id,
+    String name,
+    LocalDate date,
+    LocalTime time
+) {
+
+    public static Reservation of(Long id, Reservation reservation) {
+        return new Reservation(id, reservation.name, reservation.date, reservation.time);
+    }
+}
